agent-backend/requirements.txt
groq

# LangChain Stack (latest compatible versions)
langchain
langchain-core
langchain-community

# LangGraph
langgraph

# LLM Integrations
langchain-groq
# API Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
mcp